helm: do not terminate absent instance

+++ contrail-helm/aws/cleanup.sh
@@ -11,16 +11,20 @@ source $ENV_FILE
 
 errors="0"
 
-aws ec2 terminate-instances --instance-ids $instance_id
-[[ $? == 0 ]] || errors="1"
-if [[ $? == 0 ]]; then
-  timeout -s 9 120 aws ec2 wait instance-terminated --instance-ids $instance_id
-  echo "INFO: Instance terminated."
+if [[ -n "$instance_id" ]] ; then
+  aws ec2 terminate-instances --instance-ids $instance_id
+  [[ $? == 0 ]] || errors="1"
+  if [[ $? == 0 ]]; then
+    timeout -s 9 120 aws ec2 wait instance-terminated --instance-ids $instance_id
+   ec ho "INFO: Instance terminated."
+  fi
 fi
 
-rm kp
-aws ec2 delete-key-pair --key-name $key_name
-[[ $? == 0 ]] || errors="1"
+if [[ -f "$WORKSPACE/kp" ]] ; then
+  rm "$WORKSPACE/kp"
+  aws ec2 delete-key-pair --key-name $key_name
+  [[ $? == 0 ]] || errors="1"
+fi
 
 aws ec2 detach-internet-gateway --internet-gateway-id $igw_id --vpc-id $vpc_id
 [[ $? == 0 ]] || errors="1"
@@ -35,4 +39,6 @@ aws ec2 delete-vpc --vpc-id $vpc_id
 
 if [ $errors == "0" ]; then
   rm $ENV_FILE
+else
+  mv $ENV_FILE $ENV_FILE.$BUILD_ID
 fi
